added CRUD tests for Repositories(shuttle,spaceliner, region, port,pad) and RepositoriesSuiteTests: ALL Passing

+++ src/main/java/com/perscholas/astrodia/models/Flight.java
@@ -30,7 +30,7 @@ public class Flight {
     Timestamp arriving;
     Integer seatsAvailable;
     @NumberFormat(style= NumberFormat.Style.CURRENCY)
-    Double pricePerSeat;
+    int pricePerSeat;
     @Column(updatable = false)
     @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "dd-MM-yyyy hh:mm:ss")
     Timestamp createdAt;
